services/dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow==10.1.0",
    "plotly==5.18.0",
    "requests==2.31.0",
    "streamlit==1.29.0",
]
